Improved key derivation for asym enc iss: #9
The payload key now uses HKDF-SHA256 with the info parameter
set to the handshake hash of the noise message in order to bind
the noise handshake to the file ciphertext.

This is a breaking change for the file format. All asymmetrically
encrypted files will need to be decrypted with the old version
and re-encrypted with the new one.